Working on Ubuntu+Nvidia
Needed to change the included header for nvidia from oclUtils.h
to opencl.h, and add the path to the FindOpenCL.cmake file.

Performance on the GTX 260c216 is somewhat lower than previous
versions. Will work on that. I believe it stems primarily from
not having multi-threaded matrix exponentiation. This does not
change the fact that 1.7 of the 6 seconds used by the algorithm
for SmallCodons.bf are in the queue. What can I do to reduce
this to less than the time taken by the CPU version? Increase the
problem size of a given kernel perhaps? Clumping the leaves into
one kernel would cut that part of the execution time in half or
more. The results kernel also takes as much time as all of the leaf
kernels combined. Cleaning this up could do some real good.