Reduced the size of many of the variables passed to the kernels.
This is in hopes of reducing per thread register consumption to
increase occupancy for the internal kernel. As there are still
more than twenty registers needed for the computation, reduing the
size of the variables will hopefully allow for more efficient
register file division.

This will need to be checked using the nvidia profiler, as the
AMD profiler does not give that level of information.