Branched to start merging relevant kernels.
Optimally you would just pass a representation of updateNodes
and everything would be taken care of on the GPU.

But Ambig nodes would be challenging. I suppose reconsolidation of
leaf and ambiguous nodes might be an option.

And accession into global memory might be a bit slow.

I'll have to think about this.