Using pinned memory decreased total execution time
however according to gDEBugger reads aren't _that_ much faster. That
being said > half the time is still spent in either reading the buffer
or in the results buffer. Minimizing this would be awesome. This could
be done by moving the rest of the pipeline to the GPU or finding some
way to speed this up.

TODO: Check it out on NVIDIA.