Started reworking the reduction kernel
Pretty much replacing the old serial reduction to powerOfTwoAbove(sites)
with a two stage parallel reduction that uses more of the work items.

The first stage reduces sites * characters to sites, the second reduces
this to just one value which needs to be returned.

So far so good, however it is not always the case that sites is a power
of two. The second part of the parallel reduction needs to be implemented
over a smaller set of the work items to get the power of two we need.
There is still some to be done with this in the kernel code.