Planned solution for multiple category ComputeBlock exections.
Catch early and pass the situation to OpenCL host code. In general
transform what OCL host code is already ther into a library of sorts,
with general functions for multiplication etc that are useful elsewhere.
This call will be ifdef'd early, passed to the host code, farmed out
to general OpenCL functions, recombined and returned.