MultiGPU working
Worked out boundary cases (making sure the number of sites was
divisible by the number GPUs and the size of the local work group).

Works fine now, but not super fast. In fact only marginally faster
than a single gpu on longcodon and slower by a third on smallcodon.

Because of the reduced size (or something) the whole thing is a bit
faster on one gpu, however.

From the profiler it appears that performance gains at the kernel
level are offset by amplified host->device device->host transfers.