Very nearly correct.
Reimplemented the results kernel so that it actually works
(node_cache) still cannot be read off the GPU, for whatever reason.
The answer is very close to being correct, it could be a scaling
issue, it could be a model reading issue, it could be a math issue,
it could be an algorithm issue.

For whatever reason the site being written to has to be off by one
in the result kernel. The corresponding off by one in the reduction
(CPU) might be the cause of the error as well.