Working on CPUs and GPUs, local mem issue
There is an issue with local memory. I am not able to allocate
more than 16kb of local memory from within the kernel. If I allocate
two 8kb buffers and leave one "empty", the results suggest that
the kernel will use both buffers for storing the data meant for just
the first buffer.