A large number of general fixes.
Mostly in the Mulmullibcpp files. They're mostly working, but
producing some incorrect answers here and there. I'm tempted to say
taht the root of the problem lies with the scaling routines. Of
course all numbers are plausible in this sutation, but it worth looking into.
In order to settle thigs I think I will copy the mulmul files back
to the general library and write a few test cases.